Add SFC feature to an ODL based openstack deployment
This commit adds Service Function Chaining(SFC) basic installation opportunity
to a Fuel orchestrated openstack cluster. SFC installation is integrated into
fuel-plugin-opendaylight, since it covers orchestration tasks only and
the used SFC is closely tied to the OpenDayLight installation used.
This is the Fuel7 uplift of the original change set.

Change-Id: Ib9b4b8fbd5964453d1a2478abaa74c21849f11d0