[SDNVPN] Change order of set manager/Use PUT
In this commit two things are changed.
The manager of ovs will be set after
all the configuration is done.
That guarantees that the config on ODL
is stabel and now resync has to be done.

In addition the Trunk Endpoint information
is not delete anymore when a new tunnel
is added.

Change-Id: I33731fad724d41665998abe69ad61c995a518910