ODL leveldb, leveldbjni: Enable arm64 support.
ODL depends on a native library (leveldb).

ODL Beryllium and Boron versions are still using a native x86 version
of this library (wrapped in leveldbjni).

There is no upstream version of leveldbjni supporting arm64 leveldb,
so we need to compile it when deploying ODL.
This is done when running odl-install puppet manifest.

For more info about this ODL limitation, see [1].

[ Alexandru.Avadanii@enea.com ]
Remove JAVA8 changes after upstream rework.
Convert puppet patching to prebuilt DEB package, as part of [2].

[1] https://bugs.opendaylight.org/show_bug.cgi?id=3973
[2] https://jira.opnfv.org/browse/ARMBAND-114

Change-Id: I66610b49a2957b584d252f49d2d6ed054c15bf9c
Signed-off-by: Florin Dumitrascu <florin.dumitrascu@enea.com>
Signed-off-by: Stanislaw Kardach <stanislaw.kardach@cavium.com>
Signed-off-by: Alexandru Avadanii <Alexandru.Avadanii@enea.com>