Add NSH Subtype patch capability
The patch should be explicitly enabled, since its build time is 1 hour
and only applies while the openvswitch in fuel-plugin-ovs is not
compatible with the actual NSH openflow implementation of OpenDayLight.

The patch's ODLTAG (can be set via environment variable) should mach with
OpenDayLight release being patched. Normally it is
- release/beryllium-sr1 for ditribution-karaf-0.4.1-Beryllium-SR1.tar.gz
- release/beryllium-sr2 for ditribution-karaf-0.4.2-Beryllium-SR2.tar.gz
- stable/beryllium      for ditribution-karaf-0.4.x-SNAPSHOT.tar.gz

There is a possibility to build the patch separately and put into
folder odl_package/nsh-pritesh-patch. Remember checking the file's naming
in this case. If the plugin encounters with the supplied jarball it won't
build itself but will try using the supplied one.

Technically the patch's jarball is produced by docker following the way
jenkins job integration-distribution-verify-beryllium of
jenkins.opendaylight.org uses.

Change-Id: Icea96a0f8879d1d1a1df235002f947ce81b79b11
Signed-off-by: Ferenc Cserepkei <ferenc.cserepkei@ericsson.com>